Change Flipslide, style nodes and comments, search block
Change FlipSlide to Mega Nav, make comment and node submitted and meta markup consistant, add card styles and start refining how Responsive menus will work in the header and leaderboard regions, also how the search will work in those regions. Restyle search to use a button by default. In short this is the first commit that starts pushing the Standard theme towards being a real usable theme, and opening the door for the AT Minimal theme to come back as a minimalist alternative.