BIG update for JS layout option
This new option allows users to select JavaScript to apply layout classes to the markup, as opposede to the standard method of using PHP/twig to generate those classes and printing them in the page template. The advantage is the layout will never break because an emtpy region prints and the wrong layout classes are appled (the JS method will never make this error).

Signed-off-by: jmburnz <jeff@adaptivethemes.com>